TDKN-53 : moved Property into property package
