+++ daikon/src/test/java/org/talend/daikon/properties/PropertiesTest.java
@@ -24,6 +24,9 @@ import org.junit.Test;
 import org.junit.rules.ErrorCollector;
 import org.talend.daikon.NamedThing;
 import org.talend.daikon.properties.presentation.Form;
+import org.talend.daikon.properties.property.Property;
+import org.talend.daikon.properties.property.PropertyFactory;
+import org.talend.daikon.properties.property.PropertyValueEvaluator;
 import org.talend.daikon.properties.test.PropertiesTestUtils;
 import org.talend.daikon.properties.testproperties.PropertiesWithDefinedI18N;
 import org.talend.daikon.properties.testproperties.TestProperties;
