+++ daikon/src/test/java/org/talend/daikon/properties/test/PropertiesTestUtils.java
@@ -28,9 +28,9 @@ import org.talend.daikon.NamedThing;
 import org.talend.daikon.properties.AnyPropertyVisitor;
 import org.talend.daikon.properties.Properties;
 import org.talend.daikon.properties.Properties.Deserialized;
-import org.talend.daikon.properties.Property;
 import org.talend.daikon.properties.presentation.Form;
 import org.talend.daikon.properties.presentation.Widget;
+import org.talend.daikon.properties.property.Property;
 import org.talend.daikon.properties.service.PropertiesService;
 
 public class PropertiesTestUtils {
