+++ daikon/src/test/java/org/talend/daikon/properties/testproperties/PropertiesWithDefinedI18N.java
@@ -12,12 +12,12 @@
 // ============================================================================
 package org.talend.daikon.properties.testproperties;
 
-import static org.talend.daikon.properties.PropertyFactory.*;
+import static org.talend.daikon.properties.property.PropertyFactory.*;
 
 import org.talend.daikon.i18n.GlobalI18N;
 import org.talend.daikon.i18n.I18nMessages;
 import org.talend.daikon.properties.Properties;
-import org.talend.daikon.properties.Property;
+import org.talend.daikon.properties.property.Property;
 
 public class PropertiesWithDefinedI18N extends Properties {
