+++ daikon/src/test/java/org/talend/daikon/properties/testproperties/nestedprop/NestedProperties.java
@@ -12,11 +12,11 @@
 // ============================================================================
 package org.talend.daikon.properties.testproperties.nestedprop;
 
-import static org.talend.daikon.properties.PropertyFactory.*;
+import static org.talend.daikon.properties.property.PropertyFactory.*;
 
 import org.talend.daikon.properties.Properties;
-import org.talend.daikon.properties.Property;
 import org.talend.daikon.properties.presentation.Form;
+import org.talend.daikon.properties.property.Property;
 
 public class NestedProperties extends Properties {
